solve compile warnings.

+++ src/type/callable.h
@@ -11,6 +11,6 @@ class Type;
 class RivCallable
 {
 public:
-	virtual int arity() = 0;
+	virtual size_t arity() = 0;
 	virtual Type call(Interpreter& interpreter, const std::vector<Type>& arguments) = 0;
 };
